Reworks the "Reformat FASTA" command and dialog
The "Reformat FASTA" command has been reworked, mainly, to make it
easier to understand and manage by the users. The associated dialog now
simplifies the reformat process by showing a preview of the reformatting
result of the first sequence and by including some advanced controls to
select the final sequence name format.