Fixes a bug causing GetORF with remove new lines to fail
The option to remove new lines in sequences of the GetORF operation was
causing the operation to fail, as it was using a negative value for the
fragment length, which is no longer supported since v0.18.0.

This commit solves this problem by using the
SequenceLengthConfiguration.buildRemoveLineBreaks() method instead of
the SequenceLengthConfiguration.buildChangeFragmentLength(int) method
currently used.