Adds support for recent BLAST versions
The BLAST operations were failing with recent BLAST versions (tested
with 2.2.31+ version), as the output format has changed. The name of the
sequences in the output file is no longer prefixed with "lcl|", which
was used by BDBM as a marker to retrieve the sequences.

With this commit, sequences are now identified by the sequence start
symbol (">") instead of the prefix. This make the operation compatible
with a wider range of BLAST versions.