Removes an unnecessary code form ReformatFastaCommandDialog
The "Renaming Mode" combobox was launching a double asynchronous dialog
pack that was unnecessary and that may be causing some repainting
problems. This commit removes the duplicated call.

In addition, a "revalidate" invocation was added to the asynchronous
dialog pack, in order to force a dialog size recalculation before
packing.