Fixes unstable Splign/Compart behaviour
The first step of the Splign/Compart pipeline is the creation of
a bidirectional genome containing the sequences of original
genome and the reverse-complement counterparts of these sequences.
So far, this pipeline created two intermediante FASTA files
containing the reverse-complement sequences, the renamed
reverse-complement sequences, and the third and final file
containing the bidirectional genome. We have observed that running
Splign in a directory with other FASTA files than the two
required inputs (the bidirectional genome and the CDS) makes it
to fail to annotate genes in some environments while behaving
properly on others. It seems that the presence of these
additional FASTA files is a problem to Splign under some
circunstances and since these files are not required they are
now removed from the temporary directory created to run
the Splign/Compart pipeline.